Fix crash previewing faulty template Closes #3815
